+++ crates/adapters/src/format/parquet/mod.rs
@@ -4,7 +4,9 @@ use std::{borrow::Cow, sync::Arc};
 
 use actix_web::HttpRequest;
 use anyhow::{bail, Result as AnyResult};
-use arrow::datatypes::{DataType, Field, Schema, TimeUnit};
+use arrow::datatypes::{
+    DataType, Field as ArrowField, Fields, IntervalUnit as ArrowIntervalUnit, Schema, TimeUnit,
+};
 use arrow::record_batch::RecordBatch;
 use bytes::Bytes;
 use erased_serde::Serialize as ErasedSerialize;
@@ -26,7 +28,7 @@ use crate::{
 };
 use pipeline_types::format::json::JsonFlavor;
 use pipeline_types::format::parquet::{ParquetEncoderConfig, ParquetParserConfig};
-use pipeline_types::program_schema::{ColumnType, Relation, SqlType};
+use pipeline_types::program_schema::{ColumnType, Field, IntervalUnit, Relation, SqlType};
 
 #[cfg(test)]
 mod test;
@@ -223,6 +225,23 @@ impl OutputFormat for ParquetOutputFormat {
 }
 
 fn relation_to_parquet_schema(relation: &Relation) -> Result<SerdeArrowSchema, ControllerError> {
+    fn field_to_arrow_field(f: &Field) -> ArrowField {
+        ArrowField::new(
+            &f.name,
+            columntype_to_datatype(&f.columntype),
+            f.columntype.nullable,
+        )
+    }
+
+    fn struct_to_arrow_fields(fields: &[Field]) -> Fields {
+        Fields::from(
+            fields
+                .iter()
+                .map(field_to_arrow_field)
+                .collect::<Vec<ArrowField>>(),
+        )
+    }
+
     // The type conversion is chosen in accordance with our internal
     // data types (see sqllib). This may need to be adjusted in the future
     // or made configurable.
@@ -244,8 +263,21 @@ fn relation_to_parquet_schema(relation: &Relation) -> Result<SerdeArrowSchema, C
             SqlType::Timestamp => DataType::Timestamp(TimeUnit::Millisecond, None),
             SqlType::Date => DataType::Date32,
             SqlType::Null => DataType::Null,
-            SqlType::Binary | SqlType::Varbinary | SqlType::Interval => todo!(),
-            SqlType::Array => todo!("handle array types"),
+            SqlType::Binary => DataType::LargeBinary,
+            SqlType::Varbinary => DataType::LargeBinary,
+            SqlType::Interval(
+                IntervalUnit::YearToMonth | IntervalUnit::Year | IntervalUnit::Month,
+            ) => DataType::Interval(ArrowIntervalUnit::YearMonth),
+            SqlType::Interval(_) => DataType::Interval(ArrowIntervalUnit::DayTime),
+            SqlType::Array => {
+                // SqlType::Array implies c.component.is_some()
+                let array_component = c.component.as_ref().unwrap();
+                DataType::LargeList(Arc::new(ArrowField::new_list_field(
+                    columntype_to_datatype(array_component),
+                    c.nullable,
+                )))
+            }
+            SqlType::Struct => DataType::Struct(struct_to_arrow_fields(c.fields.as_ref().unwrap())),
         }
     }
 
@@ -253,13 +285,13 @@ fn relation_to_parquet_schema(relation: &Relation) -> Result<SerdeArrowSchema, C
         .fields
         .iter()
         .map(|f| {
-            Field::new(
+            ArrowField::new(
                 &f.name,
                 columntype_to_datatype(&f.columntype),
                 f.columntype.nullable,
             )
         })
-        .collect::<Vec<Field>>();
+        .collect::<Vec<ArrowField>>();
 
     SerdeArrowSchema::from_arrow_fields(&fields).map_err(|e| ControllerError::SchemaParseError {
         error: format!("Unable to convert schema to parquet/arrow: {e}"),